# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "測試用"
  },
  "servers": [
    {
      "url": "http://localhost:9527/"
    }
  ],
  "tags": [
    {
      "name": "test",
      "description": "測試用 API"
    },
    {
      "name": "auth",
      "description": "登入相關 API"
    }
  ],
  "paths": {
    "/api/test/": {
      "get": {
        "tags": [
          "test"
        ],
        "summary": "測試連線",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test/testToken": {
      "get": {
        "tags": [
          "test"
        ],
        "summary": "測試驗證TOKEN功能",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/test/test/sqlConnect": {
      "get": {
        "tags": [
          "test"
        ],
        "summary": "驗證SQL連線功能",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "登入功能",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "account": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            },
            "description": "輸入帳號",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": false
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "輸入 Token ， 方便後續功能測試"
      }
    }
  }
}